+++ desktop/src/main/java/bisq/desktop/components/controls/validator/deprecated/PriceValidator.java
@@ -15,16 +15,17 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.common.validation;
+package bisq.desktop.components.controls.validator.deprecated;
 
 import bisq.i18n.Res;
 import lombok.extern.slf4j.Slf4j;
 
 // todo better use another validation framework as in bisq 1
 @Slf4j
-public class MonetaryValidator extends InputValidator {
+@Deprecated
+public class PriceValidator extends InputValidator {
     public ValidationResult validate(String value) {
-      /*  ValidationResult result = super.validate(value);
+       /* ValidationResult result = super.validate(value);
         if (!result.isValid) {
             return result;
         }*/
@@ -32,7 +33,6 @@ public class MonetaryValidator extends InputValidator {
             return new ValidationResult(true);
         }
         try {
-            value = value.replace(",", ".");
             Double.parseDouble(value);
             return new ValidationResult(true);
         } catch (Throwable error) {
