+++ desktop/src/main/java/bisq/desktop/main/content/bisq_easy/components/AmountInput.java
@@ -19,17 +19,10 @@ package bisq.desktop.main.content.bisq_easy.components;
 
 import bisq.common.currency.Market;
 import bisq.common.monetary.Monetary;
-import bisq.desktop.common.validation.MonetaryValidator;
+import bisq.desktop.components.controls.validator.deprecated.MonetaryValidator;
 import bisq.presentation.formatters.AmountFormatter;
 import bisq.presentation.parser.AmountParser;
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.ReadOnlyBooleanProperty;
-import javafx.beans.property.ReadOnlyObjectProperty;
-import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleObjectProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
+import javafx.beans.property.*;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.scene.control.Label;
