+++ desktop/src/main/java/bisq/desktop/main/content/bisq_easy/components/PriceInput.java
@@ -22,8 +22,8 @@ import bisq.common.currency.Market;
 import bisq.common.monetary.PriceQuote;
 import bisq.common.observable.Pin;
 import bisq.desktop.common.threading.UIThread;
-import bisq.desktop.common.validation.PriceValidator;
 import bisq.desktop.components.controls.MaterialTextField;
+import bisq.desktop.components.controls.validator.deprecated.PriceValidator;
 import bisq.i18n.Res;
 import bisq.presentation.formatters.PriceFormatter;
 import bisq.presentation.parser.PriceParser;
